+++ src/views/home/HomeView.scss
@@ -18,178 +18,6 @@
   gap: 48px;
 }
 
-// ── 섹션 공통 ─────────────────────────────────────────────
-.dash-section {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.section-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.section-title {
-  display: flex;
-  align-items: center;
-  gap: 9px;
-  font-size: 15px;
-  font-weight: 700;
-  color: var(--c-text-2);
-  font-family: system-ui, sans-serif;
-  letter-spacing: -0.01em;
-}
-
-.section-icon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 28px;
-  height: 28px;
-  border-radius: 7px;
-  background: rgba(170, 59, 255, 0.12);
-  color: #c084fc;
-  flex-shrink: 0;
-}
-
-.section-action {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  padding: 6px 13px;
-  border-radius: 8px;
-  border: 1px solid rgba(170, 59, 255, 0.35);
-  background: rgba(170, 59, 255, 0.1);
-  color: #c084fc;
-  font-size: 12px;
-  font-weight: 600;
-  font-family: system-ui, sans-serif;
-  text-decoration: none;
-  transition: all 0.15s;
-
-  &:hover {
-    background: rgba(170, 59, 255, 0.2);
-    border-color: rgba(170, 59, 255, 0.6);
-    color: #d8a4ff;
-  }
-}
-
-// ── 상태 메시지 ───────────────────────────────────────────
-.state-msg {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-  padding: 40px 0;
-  font-size: 13px;
-  color: var(--c-text-faint);
-  font-family: system-ui, sans-serif;
-
-  &--error { color: #f87171; }
-}
-
-.empty-link {
-  font-size: 13px;
-  color: #c084fc;
-  text-decoration: none;
-
-  &:hover { text-decoration: underline; }
-}
-
-// ── 맵 그리드 ─────────────────────────────────────────────
-.map-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(240px, 1fr));
-  gap: 10px;
-}
-
-.map-card {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: 11px 12px;
-  border-radius: 10px;
-  border: 1px solid var(--c-border);
-  background: var(--c-surface-raised);
-  text-decoration: none;
-  transition: all 0.15s;
-  cursor: pointer;
-
-  &:hover {
-    border-color: rgba(170, 59, 255, 0.3);
-    background: rgba(170, 59, 255, 0.05);
-
-    .map-edit-icon { opacity: 1; }
-  }
-}
-
-.map-thumb {
-  width: 56px;
-  height: 42px;
-  border-radius: 5px;
-  overflow: hidden;
-  flex-shrink: 0;
-  background: var(--c-overlay);
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-    display: block;
-  }
-}
-
-.map-thumb-empty {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.map-info {
-  flex: 1;
-  min-width: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 3px;
-}
-
-.map-name {
-  font-size: 13px;
-  font-weight: 600;
-  color: var(--c-text);
-  font-family: system-ui, sans-serif;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.map-meta {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  font-size: 11px;
-  color: var(--c-text-muted);
-  font-family: system-ui, sans-serif;
-
-  span + span::before {
-    content: '·';
-    margin-right: 6px;
-  }
-}
-
-.map-tileset { font-weight: 500; }
-
-.map-edit-icon {
-  flex-shrink: 0;
-  color: var(--c-icon);
-  opacity: 0;
-  transition: opacity 0.15s;
-}
-
 // ── 네비게이션 카드 그리드 ─────────────────────────────────
 .nav-grid {
   display: grid;
